+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v0.3.0-dev
+
+* Features:
+  * Raises exception if relationship is not pre-fetched.
+  * Adds JaSerializer.EctoErrorSerializer
+
 ## v0.2.0
 
 * Features: